Move system:init observation tests to cross-provider module
The rpc-4 matrix slot runs with mock_sdk=true (dev proxy) and no real
credentials. The cross-provider module runs with mock_sdk=false and has
both MINIMAX_API_KEY and GLM_API_KEY available.

Move the system:init observation tests from rpc-model-switching.test.ts
to cross-provider-model-switch.test.ts so they run in the correct CI
slot with real credentials. Also clean up unused imports in rpc-model-switching.test.ts.
